core/control-plane/main.py
import os
from typing import List, Optional
from fastapi import FastAPI, Depends, HTTPException, status, Request
from pydantic import BaseModel
from sqlalchemy import create_engine, Column, String, Integer, Boolean, ForeignKey, JSON
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy.orm import sessionmaker, Session
from jose import jwt
import requests

# Configuration
DATABASE_URL = os.getenv("DATABASE_URL", "postgresql://admin:password@db:5432/saas_suite")
KEYCLOAK_URL = os.getenv("KEYCLOAK_URL", "http://keycloak:8080")
JWT_SECRET = os.getenv("JWT_SECRET", "super-secret-suite-key")
ALGORITHM = "HS256"

app = FastAPI(title="SaaS Suite Control Plane")

# Database setup
engine = create_engine(DATABASE_URL)
SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
Base = declarative_base()

# Models (Minimal for now, maps to init.sql)
class Tenant(Base):
    __tablename__ = "tenants"
    __table_args__ = {"schema": "control_plane"}
    id = Column(String, primary_key=True)
    name = Column(String)
    status = Column(String)

# Dependency
def get_db():
    db = SessionLocal()
    try:
        yield db
    finally:
        db.close()

# Schemas
class TokenExchangeRequest(BaseModel):
    id_token: str

class SuiteTokenResponse(BaseModel):
    access_token: str
    refresh_token: Optional[str] = None
    token_type: str = "bearer"

@app.get("/health")
def health_check():
    return {"status": "ok"}

@app.post("/auth/exchange", response_model=SuiteTokenResponse)
async def exchange_token(req: TokenExchangeRequest, db: Session = Depends(get_db)):
    """
    Exchanges a Keycloak ID Token for a Suite JWT.
    In a real implementation, this would:
    1. Verify ID Token with Keycloak.
    2. Lookup User in control_plane.users.
    3. Determine User's default tenant and roles.
    4. Sign a new Suite JWT.
    """
    # For MVP verification, we'll return a mock token
    # In full implementation, verify against Keycloak JWKS
    payload = {
        "sub": "user-123",
        "email": "admin@example.com",
        "tenant_id": "00000000-0000-0000-0000-000000000000",
        "roles": ["platform_super_admin"],
        "permissions": ["portal.tenants.manage", "portal.users.manage"]
    }
    token = jwt.encode(payload, JWT_SECRET, algorithm=ALGORITHM)
    return {"access_token": token}

@app.get("/tenants")
def list_tenants(db: Session = Depends(get_db)):
    # Query from control_plane.tenants
    return []

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)
