core/control-plane/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
requests==2.31.0
alembic==1.12.1
pydantic-settings==2.1.0
python-keycloak==3.3.0
